pyproject.toml
[project]
name = "haproxy-template-ic"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohttp-retry>=2.9.1",
    "click>=8.2.1",
    "deepdiff>=8.6.0",
    "deprecated>=1.2.18",
    "httpx>=0.28.1",
    "jinja2>=3.1.0",
    "kopf[dev]>=1.38.0",
    "kr8s>=0.20.9",
    "kubernetes>=33.1.0",
    "lazy-imports>=1.0.1",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-jaeger>=1.21.0",
    "opentelemetry-instrumentation-asyncio>=0.57b0",
    "opentelemetry-instrumentation-httpx>=0.57b0",
    "opentelemetry-sdk>=1.36.0",
    "prometheus-async>=25.1.0",
    "prometheus-client>=0.22.1",
    "pydantic>=2.0.0", # Required by generated client
    "structlog>=25.4.0",
    "tenacity>=9.1.0",
    "python-dateutil>=2.8.2", # Required by generated client
    "pyyaml>=6.0.2",
    "typing-extensions>=4.0.0", # Required by generated client
    "urllib3>=2.0.0", # Required by generated client (with asyncio support)
    "uvloop>=0.21.0",
]

[dependency-groups]
dev = [
    "bandit[toml]>=1.7.0",
    "black>=25.1.0",
    "commitizen>=4.8.3",
    "cryptography>=45.0.6",
    "deptry>=0.12.0",
    "docker>=7.1.0",
    "filelock>=3.19.1",
    "lint>=1.2.1",
    "mypy>=1.8.0",
    "openapi-python-client>=0.25.3",
    "oscrypto @ git+https://github.com/wbond/oscrypto.git@d5f3437", # Fix for OpenSSL 3.0.10+ version detection bug
    "prometheus-client>=0.22.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-kind>=22.11.1",
    "pytest-profiling>=1.8.1",
    "pytest-shared-session-scope>=0.5.0",
    "pytest-xdist>=3.8.0",
    "python-on-whales>=0.78.0",
    "requests>=2.32.4",
    "ruff>=0.12.7",
    "types-click>=7.1.8",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.32.4.20250809",
]

[project.scripts]
haproxy-template-ic = "haproxy_template_ic.__main__:cli"

[tool.uv]
package = true

[tool.uv.workspace]
members = [
    "haproxy-template-ic",
]

[tool.setuptools.packages.find]
where = [".", "codegen/haproxy_dataplane_v3"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:coroutine.*was never awaited:RuntimeWarning",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cluster-name=haproxy-template-ic-test",
    "--kubeconfig=.pytest-kind/haproxy-template-ic-test/kubeconfig",
]
markers = [
    "acceptance: marks tests as acceptance/end-to-end tests (deselect with '-m \"not acceptance\"')",
    "integration: marks tests as integration tests using real external dependencies (deselect with '-m \"not integration\"')",
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
    "unit: marks tests as unit tests with no external dependencies (lightweight fixtures only)",
]

[tool.coverage.run]
source = ["haproxy_template_ic"]
omit = [
    "tests/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "codegen/*",  # Exclude generated code
]

[tool.coverage.paths]
source = [
    "haproxy_template_ic",
    "/app/lib/python*/site-packages/haproxy_template_ic",
]

[tool.coverage.report]
omit = [
    "tests/*",
    "*/tests/*",
    "codegen/*",  # Exclude generated code
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.13"
strict = false
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = false
warn_unreachable = true

# Exclude test files and generated code for now to focus on production code
exclude = [
    "tests/.*",
    "codegen/.*",  # Exclude generated code
]

# For third-party libraries without type stubs or py.typed markers
# We only ignore imports for libraries that genuinely lack typing support
[[tool.mypy.overrides]]
module = [
    "kopf.*",        # Kubernetes operator framework - no official type stubs
    "kr8s.*",        # Kubernetes library - no official type stubs  
    "kubernetes.*",  # Official Kubernetes client - no official type stubs
    "deepdiff.*",    # Deep difference library - no official type stubs
    "docker.*",      # Docker library - no official type stubs
    "python_on_whales.*",  # Docker wrapper - no official type stubs
    "haproxy_dataplane_v3.*",  # Generated HAProxy Dataplane API client
    "circuitbreaker.*",  # Circuit breaker library - no official type stubs
]
ignore_missing_imports = true

# Be more lenient with kopf-related issues
[[tool.mypy.overrides]]
module = "haproxy_template_ic.operator"
warn_return_any = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
# Ignore kopf signature incompatibilities
disable_error_code = ["arg-type"]

[tool.bandit]
exclude_dirs = ["tests", "codegen"]  # Don't security-scan generated code

[tool.deptry]
extend_exclude = ["tests", "coverage_wrapper.py", "codegen"]  # Test files and generated code can use dev dependencies
per_rule_ignores = { "DEP002" = ["deprecated", "aiohttp-retry", "pydantic", "python-dateutil", "typing-extensions", "urllib3", "lazy-imports", "httpx"], "DEP003" = ["haproxy_dataplane_v3"] }

[tool.ruff]
exclude = [
    "codegen",  # Don't lint generated code
]

[tool.ruff.lint]
extend-ignore = [
    # Allow generated code patterns that don't follow our rules
]

[tool.commitizen]
version = "0.1.0"
version_files = ["pyproject.toml:version"]
tag_format = "v$version"
update_changelog_on_bump = true
major_version_zero = true
allowed_prefixes = ["feat", "fix", "docs", "style", "refactor", "perf", "test", "chore", "ci", "build", "revert"]
